+++ src/components/mod.rs
@@ -3,7 +3,7 @@ pub mod movement;
 pub mod object;
 pub mod player;
 pub mod position;
-pub mod selectable;
+pub mod selection;
 pub mod size;
 pub mod sprite;
 pub mod view;